refactor(merge): deduplicate file I/O helpers across format modules
Made-with: Cursor
