+++ src/merge/markdown.rs
@@ -26,9 +26,15 @@
 
 use log::warn;
 
+use super::{read_file_as_string, read_file_as_string_optional, write_string_to_file};
 use crate::config::{InsertPosition, MarkdownMergeOp};
 use crate::error::{Error, Result};
-use crate::filesystem::{File, MemoryFS};
+use crate::filesystem::MemoryFS;
+
+#[cfg(test)]
+use super::ensure_trailing_newline;
+#[cfg(test)]
+use crate::filesystem::File;
 
 /// Split content into lines while preserving trailing newline information
 ///
@@ -213,46 +219,6 @@ pub fn apply_markdown_merge_operation(fs: &mut MemoryFS, op: &MarkdownMergeOp) -
     }
 
     write_string_to_file(fs, dest_path, serialized)
-}
-
-// File I/O helpers
-
-fn read_file_as_string(fs: &MemoryFS, path: &str) -> Result<String> {
-    match fs.get_file(path) {
-        Some(file) => String::from_utf8(file.content.clone()).map_err(|_| Error::Merge {
-            operation: format!("read {}", path),
-            message: "File content is not valid UTF-8".to_string(),
-        }),
-        None => Err(Error::Merge {
-            operation: format!("read {}", path),
-            message: "File not found in filesystem. Was it possibly renamed or excluded by a preceding operation?".to_string(),
-        }),
-    }
-}
-
-fn read_file_as_string_optional(fs: &MemoryFS, path: &str) -> Result<Option<String>> {
-    if let Some(file) = fs.get_file(path) {
-        Ok(Some(String::from_utf8(file.content.clone()).map_err(
-            |_| Error::Merge {
-                operation: format!("read {}", path),
-                message: "File content is not valid UTF-8".to_string(),
-            },
-        )?))
-    } else {
-        Ok(None)
-    }
-}
-
-fn ensure_trailing_newline(mut content: String) -> String {
-    if !content.ends_with('\n') {
-        content.push('\n');
-    }
-    content
-}
-
-fn write_string_to_file(fs: &mut MemoryFS, path: &str, content: String) -> Result<()> {
-    let normalized = ensure_trailing_newline(content);
-    fs.add_file(path, File::from_string(&normalized))
 }
 
 #[cfg(test)]
